+++ starbridge-apps/sealed-auction/tests/deos_seam.rs
@@ -289,7 +289,7 @@ fn the_executor_re_enforces_a_phase_rewind_is_refused_strictmonotonic() {
     // StrictMonotonic clause bites).
     let rewind = vec![Effect::SetField {
         cell,
-        index: PHASE_SLOT,
+        index: PHASE_SLOT as u64,
         value: field_from_u64(PHASE_COMMIT),
     }];
     let action = cclerk.make_action(cell, "resolve", rewind);
@@ -307,7 +307,7 @@ fn the_executor_re_enforces_a_phase_rewind_is_refused_strictmonotonic() {
     // STALL: a resolve that leaves PHASE at REVEAL (no advance) — strict bites.
     let stall = vec![Effect::SetField {
         cell,
-        index: PHASE_SLOT,
+        index: PHASE_SLOT as u64,
         value: field_from_u64(PHASE_REVEAL),
     }];
     let action = cclerk.make_action(cell, "resolve", stall);
